Native iOS: library-first builder with a reorderable queue tray
Redesign the setlist builder (your direction): browse the library inline up
top, tap to add, and the setlist forms in a reorderable queue tray at the
bottom — like building a playlist. This folds the separate "Add items" picker
sheet back into one screen and replaces swipe-to-delete with the queue's
always-on drag grips + an inline remove, which dissolves the swipe-vs-gap
tension entirely.

- Library section: the full browse (search / type pills / sort / tag) over the
  spaced Library cards with the add/added toggle.
- Queue tray ("In this session"): compact rows with an inline remove and an
  always-on reorder grip (forced edit mode); reorderSetlist on move; caps height
  and scrolls.
- Deletes the now-redundant SessionItemPickerSheet + SetlistEntryRow (replaced by
  LibraryItemCard up top and the new SetlistQueueRow below) and their tests.

UITest rewritten for the new flow (add via library cards, remove via the queue);
builder snapshots re-recorded.
